+++ tests/test_cache_wrapper.py
@@ -1,37 +1,12 @@
 import unittest
 import mlx.core as mx
-import mlx.nn as nn
 from mlx_engine.cache_wrapper import CacheWrapper
 from mlx_engine.cache import ShiftingKVCache
+from tests.test_cache_generic import TestCache
 from tests.utils import DummyModel
 
 
-class TestCacheWrapper(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        """Set up test resources that will be shared across all test methods"""
-        cls.kv_head_dim = 4
-        cls.bsz = 1
-        cls.n_kv_heads = 1
-        # cannot be used raw: must be wrapped in the cache.rope workaround impl
-        cls._rope = nn.RoPE(
-            dims=cls.kv_head_dim, traditional=False, base=100000, scale=1.0
-        )
-
-    @classmethod
-    def make_random_kv(cls, seqlen: int):
-        """Helper method to make a random key/value tensor of the right shape"""
-        return mx.random.normal(
-            (cls.bsz, cls.n_kv_heads, seqlen, cls.kv_head_dim),
-            scale=1.0,
-            dtype=mx.float32,
-        )
-
-    def assertArrEqual(self, a: mx.array, b: mx.array):
-        """Assert that two tensors are equal over the sequence length dimension"""
-        self.assertEqual(a.shape, b.shape)
-        self.assertTrue(mx.allclose(a, b), "Tensors are not equal")
-
+class TestCacheWrapper(TestCache):
     def test_find_matching_sequence_length_with_mismatch(self):
         """Test when there's a mismatch in the tokens"""
         # Create two arrays with a known common prefix [1, 2, 3]
